+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Added `jaeger configuration slew` command.
 * Added chiller control.
 * A keyword `folded` is updated after a trajectory indicating if the array is folded. In `jaeger unwind`, if the array is already folded (with a tolerance of 1 degree), it does nothing. `jaeger unwind --status` only reports the folded status.
+* Report `alive_at` every 60 second.
 
 ### ✨ Improved